feat: break down types to files
