+++ lib/definition.ts
@@ -7,6 +7,7 @@ import {
   BulkCreateOutputType,
   BulkRemoveOutputType,
   FilterType,
+  BulkUpdateOutputType,
 } from './type-functions';
 import { MetaKey, Metadata } from './metadata';
 import { AllowedApiType } from './shared';
@@ -24,6 +25,7 @@ export function Definition(options: DefinitionOptions = { allowedApis: 'essentia
     UpdateInputType(target);
     OutputType(target);
     BulkCreateOutputType(target);
+    BulkUpdateOutputType(target);
     BulkRemoveOutputType(target);
     PaginatedOutputType(target);
     FilterType(target);
