+++ lib/type-functions/paginate-output-type.ts
@@ -1,0 +1,49 @@
+import * as graphql from 'graphql';
+import { ObjectType, Field } from '@nestjs/graphql';
+import { Type } from 'class-transformer';
+
+import * as util from '../util';
+import { Definition } from '../definition';
+import { cacheType } from './cache-type';
+import { OutputType } from './output-type';
+
+function getType(definition: Definition): any {
+  @ObjectType(`Paginated${util.plural(definition.name)}`)
+  class Placeholder {
+    @Field(() => [OutputType(definition)])
+    @Type(() => OutputType(definition))
+    docs: any[];
+
+    @Field(() => graphql.GraphQLInt)
+    totalDocs: number;
+
+    @Field(() => graphql.GraphQLInt)
+    page: number;
+
+    @Field(() => graphql.GraphQLInt)
+    limit: number;
+
+    @Field(() => graphql.GraphQLInt)
+    totalPages: number;
+
+    @Field(() => graphql.GraphQLBoolean)
+    hasNextPage: boolean;
+
+    @Field(() => graphql.GraphQLInt)
+    nextPage: number;
+
+    @Field(() => graphql.GraphQLBoolean)
+    hasPrevPage: boolean;
+
+    @Field(() => graphql.GraphQLInt)
+    prevPage: number;
+
+    @Field(() => graphql.GraphQLInt)
+    pagingCounter: number;
+  }
+  return Placeholder;
+}
+
+export function PaginatedOutputType(definition: Definition) {
+  return cacheType(() => getType(definition), definition, 'PaginatedOutputType');
+}
